requirements.txt
pandas
fastapi
uvicorn[standard]
google-genai
groq
python-dotenv
sqlalchemy
openpyxl
pypdf
pydicom
pillow
numpy
nibabel
scikit-learn
joblib
torch
torchvision
shap
